remove configuration screen
