+++ app/src/main/java/seamuslowry/paintracker/MainActivity.kt
@@ -23,7 +23,7 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background,
                 ) {
-                    Navigation(startDestination = Screen.Configuration.identifier)
+                    Navigation(startDestination = Screen.Entry.identifier)
                 }
             }
         }
